python-api/requirements.txt
grpcio
grpcio-tools
easyocr
ultralytics
opencv-python-headless
pillow
numpy
